services/image_processor/pyproject.toml
[project]
name = "image-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.6",
    "ipython==8.12.3",
    "ipywidgets==8.1.5",
    "matplotlib==3.8.4",
    "minio>=7.2.14",
    "numba==0.60.0",
    "numba-progress==1.1.0",
    "numpy==2.0.1",
    "opencv-python==4.10.0.84",
    "pillow==11.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "rawpy==0.22.0",
    "scipy==1.15.0",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]
